containerization/requirements.txt
requests
pathlib
pyyaml
